メイン機能の分割
